int to size_t & proccessing memory exception

+++ 01/allocator.h
@@ -1,8 +1,8 @@
 #ifndef ALLOCATOR_H_INCLUDED
 #define ALLOCATOR_H_INCLUDED
 
-void makeAllocator(int num);
-char *alloc(int num);
+void makeAllocator(size_t num);
+char *alloc(size_t num);
 void reset();
 
 #endif // ALLOCATOR_H_INCLUDED
